# lms_back/api/docs/swagger.yaml
definitions:
...
  models.CreateStudent:
    properties:
      age:
        type: integer
      email:
        type: string
      full_name:
        type: string
      group_id:
        type: string
      login:
        type: string
      paid_sum:
        type: number
      password:
        type: string
      status:
        type: string
    type: object
  models.CreateTask:
    properties:
      lesson_id:
        type: string
      score:
        type: string
      task:
        type: string
    type: object
...
  models.GetAllStudentsResponse:
    properties:
      count:
        type: integer
      students:
        items:
          $ref: '#/definitions/models.Student'
        type: array
    type: object
  models.GetAllTasksResponse:
    properties:
      count:
        type: integer
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
    type: object
...
  models.GetStudent:
    properties:
      age:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: integer
      email:
        type: string
      full_name:
        type: string
      group_id:
        type: string
      id:
        type: string
      login:
        type: string
      paid_sum:
        type: number
      password:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.GetTask:
    properties:
      created_at:
        type: string
      id:
        type: string
      lesson_id:
        type: string
      score:
        type: string
      task:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.Student:
    properties:
      age:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: integer
      email:
        type: string
      full_name:
        type: string
      group_id:
        type: string
      id:
        type: string
      login:
        type: string
      paid_sum:
        type: number
      password:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      id:
        type: string
      lesson_id:
        type: string
      score:
        type: string
      task:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateStudent:
    properties:
      age:
        type: integer
      email:
        type: string
      full_name:
        type: string
      group_id:
        type: string
      id:
        type: string
      login:
        type: string
      paid_sum:
        type: number
      password:
        type: string
      status:
        type: string
    type: object
  models.UpdateTask:
    properties:
      lesson_id:
        type: string
      score:
        type: string
      task:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /student:
    get:
      description: This API returns student list
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: limit per page
        in: query
        name: limit
        type: integer
      - description: search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllStudentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: get all students
      tags:
      - student
    post:
      consumes:
      - application/json
      description: This api is creates a new student and returns its id
      parameters:
      - description: student
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/models.CreateStudent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: create a student
      tags:
      - student
  /student/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a student by its ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: delete a student by ID
      tags:
      - student
    get:
      consumes:
      - application/json
      description: Retrieves a student by its ID
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetStudent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: return a student by ID
      tags:
      - student
    put:
      consumes:
      - application/json
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      - description: student
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStudent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: update a student
      tags:
      - student
  /task:
    get:
      description: This API returns task list
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: limit per page
        in: query
        name: limit
        type: integer
      - description: search keyword
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllTasksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: get all tasks
      tags:
      - task
    post:
      consumes:
      - application/json
      description: This api is creates a new task and returns its id
      parameters:
      - description: task
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/models.CreateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateTask'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: create a task
      tags:
      - task
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: delete a task by ID
      tags:
      - task
    get:
      consumes:
      - application/json
      description: Retrieves a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetTask'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: return a task by ID
      tags:
      - task
    put:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: update a task
      tags:
      - task
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"